P0 sprint: CI green, MobileSAM decoder, label export workflow.
Normalize YOLO confidence to [0,1], upgrade numpy 2.x for scipy, replace
defective MobileSAM decoder ONNX, add reviewed-label export scripts, raise
ROAD_SEG_CONF_THRESHOLD to 0.5, and clear all ruff violations.

+++ edgevision-mw/app/ai/mask_utils.py
@@ -31,7 +31,7 @@ def polygon_to_mask(polygon: list, width: int, height: int) -> np.ndarray:
     if not polygon or width <= 0 or height <= 0:
         return mask
     pts = np.array(
-        [[int(round(p[0] * width)), int(round(p[1] * height))] for p in polygon],
+        [[round(p[0] * width), round(p[1] * height)] for p in polygon],
         dtype=np.int32,
     )
     if pts.shape[0] >= 3:
